Rename maper package to mapper
